[PT2] Fix FakeTensor dtype for torch.unique inverse_indices and counts (#179401) (#179401)
Summary:

The FakeTensor dispatch for torch.unique (_unique2, unique_dim,
unique_consecutive) was using arg.new_empty() to create inverse_indices
and counts tensors, which inherits the input tensor's dtype. The real
C++ kernels always create these with dtype=kLong (int64). This mismatch
causes scatter() to fail with "Expected dtype int32/int64 for index"
when using inverse_indices from torch.unique during torch.export.

Test Plan:
buck test fbcode//mode/dev fbcode//caffe2/test:fake_tensor -- test_unique_output_dtype
Pass 2. Fail 0: https://www.internalfb.com/intern/testinfra/testrun/25051272939378423

arc lint: No lint issues.

Reviewed By: felixsu2006

Differential Revision: D99584353

Pull Request resolved: https://github.com/pytorch/pytorch/pull/179401
Approved by: https://github.com/tugsbayasgalan

+++ torch/_subclasses/fake_impls.py
@@ -439,15 +439,19 @@ def _unique(
 
     return_if_dim_and_cpu = dim is not None and arg.fake_device == torch.device("cpu")
     if return_inverse or return_if_dim_and_cpu:
-        inverse = arg.new_empty(arg.shape if dim is None else (arg.shape[dim],))
+        inverse = arg.new_empty(
+            arg.shape if dim is None else (arg.shape[dim],), dtype=torch.int64
+        )
     else:
-        inverse = arg.new_empty(0)
+        inverse = arg.new_empty(0, dtype=torch.int64)
     ret.append(inverse)
 
     if return_counts or return_if_dim_and_cpu:
-        counts = arg.new_empty(ret[0].shape if dim is None else (ret[0].shape[dim],))
+        counts = arg.new_empty(
+            ret[0].shape if dim is None else (ret[0].shape[dim],), dtype=torch.int64
+        )
     else:
-        counts = arg.new_empty(0)
+        counts = arg.new_empty(0, dtype=torch.int64)
     ret.append(counts)
 
     return tuple(ret)
